More x64 changes
git-svn-id: http://miranda.googlecode.com/svn/trunk@10473 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/plugins/freeimage/Miranda/main.cpp
@@ -821,7 +821,7 @@ extern "C" BOOL __declspec(dllexport) dib2mempng( BITMAPINFO* pbmi, png_byte* pD
 		png_write_end(png_ptr, info_ptr);
 
 		if ( pResultLen != NULL )
-			*pResultLen = sBuffer.mBufPtr;
+			*pResultLen = (long)sBuffer.mBufPtr;
 
 		free( pTempBuffer );
 	}
